requirements.txt
# --------- pytorch --------- #
torch>=1.10.0
torchvision>=0.11.0
pytorch-lightning>=1.5.10
torchmetrics>=0.7.0

# --------- hydra --------- #
hydra-core>=1.1.0
hydra-colorlog>=1.1.0
hydra-optuna-sweeper>=1.1.0

# --------- loggers --------- #
wandb~=0.12.19
# neptune-client
# mlflow
# comet-ml
tensorboard

# --------- linters --------- #
pre-commit      # hooks for applying linters on commit
black           # code formatting
isort           # import sorting
flake8          # code analysis
nbstripout      # remove output from jupyter notebooks

# --------- others --------- #
pyrootutils
python-dotenv~=0.20.0
protobuf~=3.19.0
rich~=12.4.4
pytest~=7.1.2
sh~=1.14.2
pudb            # debugger

seaborn~=0.11.2
omegaconf~=2.2.2
transformers~=4.19.2
packaging~=21.3
datasets~=2.2.2
beautifulsoup4~=4.9.0

matplotlib~=3.5.1



# --------- doc2json --------- #
boto3~=1.9.147
requests~=2.21.0
Flask~=1.0.2
tqdm
lxml
python-magic~=0.4.18
latex2mathml~=2.16.2

